backend/requirements.txt
fastapi
motor
pydantic
python-dotenv
google-generativeai
